Refactor API routes to use consistent naming conventions
