+++ backend/server.js
@@ -4,7 +4,7 @@ import authRoutes from "./routes/authRoutes.js";
 import profileRoutes from "./routes/ProfileRoutes.js";
 import notificationRoutes from "./routes/notificationRoutes.js";
 import { verifyToken } from "./middleWare/authMiddleware.js";
-import HrRouter from "./routes/HrRouter.js";
+
 
 const app = express();
 
@@ -20,7 +20,7 @@ app.use("/api/profile", verifyToken, profileRoutes);
 
 // Use the routes
 app.use("/api/notification", notificationRoutes);
-app.use("/api/Hr/auth", HrRouter);
+
 
 // Default route
 app.get("/", (req, res) => {
