+++ frontend/src/pages/Employee_Information_Management.jsx
@@ -48,7 +48,7 @@ const EmployeeInfoManagement = () => {
 
   useEffect(() => {
     axios
-      .get("http://localhost:8800/api/Hr/auth/employees")
+      .get("http://localhost:8800/api/auth/Hr/employees")
       .then((response) => {
         setEmployees(response.data);
         setFilterEmployees(response.data); // Update filterEmployees when employees are fetched
@@ -64,7 +64,7 @@ const EmployeeInfoManagement = () => {
     const fetchDepartments = async () => {
       try {
         const response = await axios.get(
-          "http://localhost:8800/api/Hr/auth/departments"
+          "http://localhost:8800/api/auth/Hr/departments"
         );
         setdepartments(departments.concat(response.data));
         //console.log(response.data);
@@ -81,7 +81,7 @@ const EmployeeInfoManagement = () => {
     const fetchJobTitles = async () => {
       try {
         const response = await axios.get(
-          "http://localhost:8800/api/Hr/auth/JobTitles"
+          "http://localhost:8800/api/auth/Hr/JobTitles"
         );
         setJobTitles(JobTitles.concat(response.data));
         //console.log(response.data);
